+++ app/Http/Controllers/UserController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\UserStatus;
 use Illuminate\Http\Request;
 use App\Models\User;
 
@@ -9,27 +10,43 @@ class UserController extends Controller
 {
     public function index()
     {
-        $users = User::all();
-        return view('admin.users', compact('users'));
+        $users = User::where('status_id', '!=', UserStatus::ADMIN)
+            ->latest()
+            ->orderBy('status', 'asc')
+            ->paginate(30);
+        return view('admin.users', [
+            'users' => $users,
+        ]);
     }
 
     public function update(Request $request, $id)
     {
-        // TODO: Update hanya untuk ubah status
-        $validatedData = $request->validate([
-        ]);
+        $user = User::find($id);
+        
+        if (!$user) {
+            return redirect()->route('admin.users.index')
+                ->with('error', 'Pengguna tidak ditemukan');
+        }
 
-        $user = User::findOrFail($id);
-        $user->update($validatedData);
+        $user->status_id = $request->status_id;
+        $user->save();
 
-        return redirect()->route('admin.users.index');
+        return redirect()->route('admin.users.index')
+            ->with('message', 'Status pengguna berhasil diubah');
     }
 
     public function destroy($id)
     {
-        $user = User::findOrFail($id);
+        $user = User::find($id);
+        
+        if (!$user) {
+            return redirect()->route('admin.users.index')
+                ->with('error', 'Pengguna tidak ditemukan');
+        }
+
         $user->delete();
 
-        return redirect()->route('admin.users.index');
+        return redirect()->route('admin.users.index')
+            ->with('message', 'Pengguna berhasil dihapus');
     }
 }
